+++ src/main/kotlin/org/vitrivr/cottontail/database/index/va/bounds/L2Bounds.kt
@@ -1,29 +1,27 @@
-package org.vitrivr.cottontail.database.index.va.lookup
+package org.vitrivr.cottontail.database.index.va.bounds
 
 import org.vitrivr.cottontail.database.index.va.signature.Marks
 import org.vitrivr.cottontail.database.index.va.signature.Signature
 import org.vitrivr.cottontail.model.values.types.RealVectorValue
 import kotlin.math.max
 import kotlin.math.pow
+import kotlin.math.sqrt
 
 /**
- * A [Bounds] implementation for Lp distance.
+ * A [Bounds] implementation for L2 (Euclidean) distance.
  *
  * @author Ralph Gasser
  * @version 1.0.0
  */
-class LpBounds(query: RealVectorValue<*>, marks: Marks, p: Int) : Bounds {
+class L2Bounds(query: RealVectorValue<*>, marks: Marks) : Bounds {
 
-    /** Lower bound of this [LpBounds]. */
+    /** Lower bound of this [L2Bounds]. */
     override var lb = 0.0
         private set
 
-    /** Upper bound of this [LpBounds]. */
+    /** Upper bound of this [L2Bounds]. */
     override var ub = 0.0
         private set
-
-    /** Exponent used for p-th root calculation. */
-    private val exp = 1.0 / p
 
     /** Cells for the query [RealVectorValue]. */
     private val rq = marks.getCells(query)
@@ -32,7 +30,7 @@ class LpBounds(query: RealVectorValue<*>, marks: Marks, p: Int) : Bounds {
     private val lat = Array(marks.marks.size) { j ->
         val qj = query[j].value.toDouble()
         Array(marks.marks[j].size) { m ->
-            doubleArrayOf((qj - marks.marks[j][m]).pow(p), (marks.marks[j][m] - qj).pow(p))
+            doubleArrayOf((qj - marks.marks[j][m]).pow(2), (marks.marks[j][m] - qj).pow(2))
         }
     }
 
@@ -60,7 +58,7 @@ class LpBounds(query: RealVectorValue<*>, marks: Marks, p: Int) : Bounds {
                 }
             }
         }
-        this.lb = lb.pow(this.exp)
-        this.ub = ub.pow(this.exp)
+        this.lb = sqrt(lb)
+        this.ub = sqrt(ub)
     }
 }
